+++ src/index.scss
@@ -380,7 +380,7 @@
 @import 'courseware/course/content-tools/contentTools.scss';
 @import 'course-home/dates-tab/Badge.scss';
 @import 'course-home/dates-tab/Day.scss';
-@import 'course-home/outline-tab/widgets/UpgradeCard.scss';
+@import 'generic/upgrade-card/UpgradeCard.scss';
 @import 'course-home/outline-tab/widgets/ProctoringInfoPanel.scss';
 @import 'course-home/progress-tab/course-completion/CompletionDonutChart.scss';
 @import 'course-home/progress-tab/grades/course-grade/GradeBar.scss';
